Move Player model to API app
