+++ backend/apiapp/models.py
@@ -1,3 +1,11 @@
 from django.db import models
+from django.contrib.auth.models import AbstractUser
 
-# Create your models here.
+
+class Player(AbstractUser):
+    email = models.EmailField(max_length=30, blank=False, null=False, unique=True)
+    username = models.CharField(max_length=20, blank=False, null=False, unique=True)
+    tournament_name = models.CharField(max_length=20, blank=False, null=True)
+
+    def __str__(self):
+        return self.email
