--- backend/authenticationapp/serializers.py
@@ -1,8 +1,0 @@
-from rest_framework import serializers
-from .models import Player
-
-
-class PlayerSerializer(serializers.ModelSerializer):
-    model = Player
-    fields = ['email']
-
